feat: add store pool stats
